Moves to Docker.Client namespace.
This is more precice since there might be more docker related
namespaces like Docker.Compose and the like which I'd like to keep
under the Docker top level namespace.
