+++ docker.cabal
@@ -1,6 +1,6 @@
 name:                docker
 version:             0.3.0.0
-synopsis:            Haskell wrapper for Docker Remote API
+synopsis:            An API client for docker written in Haskell
 description:         See README.md
 homepage:            https://github.com/denibertovic/docker-hs
 license:             BSD3
@@ -17,7 +17,7 @@ stability:           experimental
 library
   default-extensions:          DeriveGeneric, OverloadedStrings, ScopedTypeVariables, ExplicitForAll, MultiParamTypeClasses, FlexibleContexts, GeneralizedNewtypeDeriving, RankNTypes, DeriveFunctor, FlexibleInstances
   hs-source-dirs:      src
-  exposed-modules:     Docker, Docker.Client, Docker.Client.Types, Docker.Client.Internal, Docker.Client.Http
+  exposed-modules:     Docker.Client, Docker.Client.Api, Docker.Client.Types, Docker.Client.Internal, Docker.Client.Http
   -- other-modules:       Docker.Internal
   build-depends:       base >= 4.7 && < 5
                      , aeson
