--- src/Docker.hs
@@ -1,13 +1,0 @@
-{-# LANGUAGE ExplicitForAll      #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-
-module Docker (
-      module Docker.Client
-    , module Docker.Client.Types
-    , module Docker.Client.Http
-    ) where
-
-import           Docker.Client
-import           Docker.Client.Http
-import           Docker.Client.Types
-
